src/requirements.txt
alembic==1.13.1
fastapi==0.109.0
sqlmodel==0.0.14
SQLAlchemy==2.0.25
uvicorn==0.25.0
starlette-admin[i18n]==0.12.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
itsdangerous==2.1.2

pydantic[email]==2.5.3
pydantic-settings==2.1.0

passlib==1.7.4
python-jose==3.3.0
